Version 0.1
This is a quick and dirty program I made to block lots of IP addresses for a client that matched the sshd auth.logs file to the site containing a list of known brute force hacker's IP addresses. There is lots that can be done with this to extend and clean its functionality.